chore: fix typo, update doc links, clean sources, fix hard-coded colors troubles (#1174)
